nxp:imxrt 1060/1170 bifurcation and restructuring
